Add support for unique instances

+++ src/ChunkManager.hpp
@@ -24,16 +24,16 @@ namespace ChunkManager {
     void init();
     void cleanup();
 
-    extern std::map<std::tuple<int, int, int>, Chunk*> chunks;
+    extern std::map<std::tuple<int, int, uint64_t>, Chunk*> chunks;
 
-    void addNPC(int posX, int posY, int instanceID, int32_t id);
-    void addPlayer(int posX, int posY, int instanceID, CNSocket* sock);
-    bool removePlayer(std::tuple<int, int, int> chunkPos, CNSocket* sock);
-    bool removeNPC(std::tuple<int, int, int> chunkPos, int32_t id);
-    bool checkChunk(std::tuple<int, int, int> chunk);
-    void destroyChunk(std::tuple<int, int, int> chunkPos);
-    std::tuple<int, int, int> grabChunk(int posX, int posY, int instanceID);
-    std::vector<Chunk*> grabChunks(std::tuple<int, int, int> chunkPos);
+    void addNPC(int posX, int posY, uint64_t instanceID, int32_t id);
+    void addPlayer(int posX, int posY, uint64_t instanceID, CNSocket* sock);
+    bool removePlayer(std::tuple<int, int, uint64_t> chunkPos, CNSocket* sock);
+    bool removeNPC(std::tuple<int, int, uint64_t> chunkPos, int32_t id);
+    bool checkChunk(std::tuple<int, int, uint64_t> chunk);
+    void destroyChunk(std::tuple<int, int, uint64_t> chunkPos);
+    std::tuple<int, int, uint64_t> grabChunk(int posX, int posY, uint64_t instanceID);
+    std::vector<Chunk*> grabChunks(std::tuple<int, int, uint64_t> chunkPos);
     std::vector<Chunk*> getDeltaChunks(std::vector<Chunk*> from, std::vector<Chunk*> to);
-    bool inPopulatedChunks(int posX, int posY, int instanceID);
+    bool inPopulatedChunks(int posX, int posY, uint64_t instanceID);
 }
