+++ src/NPC.hpp
@@ -7,12 +7,12 @@ class BaseNPC {
 public:
     sNPCAppearanceData appearanceData;
     NPCClass npcClass;
-    int instanceID;
-    std::tuple<int, int, int> chunkPos;
+    uint64_t instanceID;
+    std::tuple<int, int, uint64_t> chunkPos;
     std::vector<Chunk*> currentChunks;
 
     BaseNPC() {};
-    BaseNPC(int x, int y, int z, int angle, int iID, int type, int id) {
+    BaseNPC(int x, int y, int z, int angle, uint64_t iID, int type, int id) {
         appearanceData.iX = x;
         appearanceData.iY = y;
         appearanceData.iZ = z;
@@ -27,7 +27,7 @@ public:
 
         chunkPos = std::make_tuple(0, 0, instanceID);
     };
-    BaseNPC(int x, int y, int z, int angle, int iID, int type, int id, NPCClass classType) : BaseNPC(x, y, z, angle, iID, type, id) {
+    BaseNPC(int x, int y, int z, int angle, uint64_t iID, int type, int id, NPCClass classType) : BaseNPC(x, y, z, angle, iID, type, id) {
         npcClass = classType;
     }
 };
